requirements.txt
pytest
allure-pytest
pyyaml
pytest-html
requests
pytest-bdd
boto3
selenium